🩹 Refactor: 프로필뷰 리팩토링
- headerView 리팩토링
