+++ PADO/PADO/iOS/Profile/View/OtherUserProfileView.swift
@@ -47,8 +47,18 @@ struct OtherUserProfileView: View {
     var body: some View {
         ZStack {
             ScrollView(.vertical, showsIndicators: false) {
-                headerView()
+                ProfileHeaderView(touchBackImage: $touchBackImage,
+                                  touchProfileImage: $touchProfileImage,
+                                  followerActive: $followerActive,
+                                  followingActive: $followingActive,
+                                  position: $position,
+                                  buttonOnOff: $buttonOnOff,
+                                  user: user,
+                                  profileVM: profileVM,
+                                  followVM: followVM,
+                                  postitVM: postitVM)
                     .padding(.top, 50)
+                
                 VStack(spacing: 0) {
                     LazyVStack(pinnedViews: [.sectionHeaders]) {
                         Section {
@@ -91,19 +101,15 @@ struct OtherUserProfileView: View {
                     HStack {
                         HStack(spacing: 0) {
                             if !user.instaAddress.isEmpty {
-                                Button {
-                                    profileVM.openSocialMediaApp(urlScheme: "instagram://user?username=\(user.instaAddress)", fallbackURL: "https://instagram.com/\(user.instaAddress)")
-                                } label: {
-                                    Image("instagramicon")
-                                }
+                                SocialMediaButton(platformName: "instagramicon",
+                                                  urlScheme: "instagram://user?username=\(user.instaAddress)",
+                                                  fallbackURL: "https://instagram.com/\(user.instaAddress)")
                             }
                             
                             if !user.tiktokAddress.isEmpty {
-                                Button {
-                                    profileVM.openSocialMediaApp(urlScheme: "tiktok://user?username=\(user.tiktokAddress)", fallbackURL: "https://www.tiktok.com/@\(user.tiktokAddress)")
-                                } label: {
-                                    Image("tiktokicon")
-                                }
+                                SocialMediaButton(platformName: "tiktokicon",
+                                                  urlScheme: "tiktok://user?username=\(user.tiktokAddress)",
+                                                  fallbackURL: "https://www.tiktok.com/@\(user.tiktokAddress)")
                             }
                             
                             if user.nameID != userNameID {
@@ -129,69 +135,19 @@ struct OtherUserProfileView: View {
             // 뒷배경 조건문
             if touchBackImage {
                 if let backProfileImageUrl = user.backProfileImageUrl {
-                    KFImage(URL(string: backProfileImageUrl))
-                        .resizable()
-                        .scaledToFit()
-                        .clipShape(RoundedRectangle(cornerRadius: isDragging ? 12 : 0))
-                        .zIndex(2)
-                        .onTapGesture {
-                            withAnimation(.interactiveSpring(response: 0.4, dampingFraction: 0.8, blendDuration: 0.8)) {
-                                self.touchBackImage = false
-                            }
-                        }
-                        .offset(position)
-                        .highPriorityGesture(
-                            DragGesture()
-                                .onChanged({ value in
-                                    self.position = value.translation
-                                    self.isDragging = true
-                                })
-                                .onEnded({ value in
-                                    withAnimation(.interactiveSpring(response: 0.4, dampingFraction: 0.8, blendDuration: 0.8)) {
-                                        if 200 < abs(self.position.height) {
-                                            self.touchBackImage = false
-                                            self.isDragging = false
-                                        } else {
-                                            self.position = .zero
-                                            self.isDragging = false
-                                        }
-                                    }
-                                })
-                        )
+                    UserProfileImageView(isTouched: $touchBackImage,
+                                         isDragging: $isDragging,
+                                         position: $position,
+                                         imageUrl: URL(string: backProfileImageUrl))
                 }
             }
             // 프로필 사진 조건문
             if touchProfileImage {
                 if let profileImageUrl = user.profileImageUrl {
-                    KFImage(URL(string: profileImageUrl))
-                        .resizable()
-                        .scaledToFit()
-                        .clipShape(RoundedRectangle(cornerRadius: isDragging ? 12 : 0))
-                        .zIndex(2)
-                        .onTapGesture {
-                            withAnimation(.interactiveSpring(response: 0.4, dampingFraction: 0.8, blendDuration: 0.8)) {
-                                self.touchProfileImage = false
-                            }
-                        }
-                        .offset(position)
-                        .highPriorityGesture(
-                            DragGesture()
-                                .onChanged({ value in
-                                    self.position = value.translation
-                                    self.isDragging = true
-                                })
-                                .onEnded({ value in
-                                    withAnimation(.interactiveSpring(response: 0.4, dampingFraction: 0.8, blendDuration: 0.8)) {
-                                        if 200 < abs(self.position.height) {
-                                            self.touchProfileImage = false
-                                            self.isDragging = false
-                                        } else {
-                                            self.position = .zero
-                                            self.isDragging = false
-                                        }
-                                    }
-                                })
-                        )
+                    UserProfileImageView(isTouched: $touchProfileImage,
+                                         isDragging: $isDragging,
+                                         position: $position,
+                                         imageUrl: URL(string: profileImageUrl))
                 }
             }
         }
@@ -223,197 +179,6 @@ struct OtherUserProfileView: View {
             profileVM.stopAllPostListeners()
             enteredNavigation = false
         }
-    }
-    
-    @ViewBuilder
-    func headerView() -> some View {
-        GeometryReader { proxy in
-            let minY = proxy.frame(in: .named("SCROLL")).minY
-            let size = proxy.size
-            let height = (size.height + minY)
-            
-            KFImage(URL(string: user.backProfileImageUrl ?? defaultBackgroundImageUrl))
-                .resizable()
-                .scaledToFill()
-                .frame(width: size.width, height: height > 0 ? height : 0, alignment: .top)
-                .overlay {
-                    ZStack(alignment: .bottom) {
-                        LinearGradient(colors: [.clear,
-                                                .main.opacity(0.1),
-                                                .main.opacity(0.3),
-                                                .main.opacity(0.5),
-                                                .main.opacity(0.8),
-                                                .main.opacity(1)],
-                                       startPoint: .top,
-                                       endPoint: .bottom)
-                        
-                        VStack(alignment: .leading, spacing: 10) {
-                            CircularImageView(size: .xxxLarge, user: user)
-                                .zIndex(touchProfileImage ? 1 : 0)
-                                .onTapGesture {
-                                    if user.profileImageUrl != nil {
-                                        position = .zero
-                                        withAnimation(.interactiveSpring(response: 0.4, dampingFraction: 0.8, blendDuration: 0.8)) {
-                                            touchProfileImage = true
-                                        }
-                                    }
-                                }
-                                .overlay {
-                                    Button {
-                                        if fetchedPostitData {
-                                            isShowingMessageView = true
-                                        }
-                                    } label: {
-                                        Circle()
-                                            .frame(width: 30)
-                                            .foregroundStyle(.clear)
-                                            .overlay {
-                                                if postitVM.messages.isEmpty {
-                                                    LottieView(animation: .named("nonePostit"))
-                                                        .paused(at: .progress(1))
-                                                        .resizable()
-                                                        .scaledToFit()
-                                                        .frame(width: 40, height: 40)
-                                                } else {
-                                                    LottieView(animation: .named("Postit"))
-                                                        .looping()
-                                                        .resizable()
-                                                        .scaledToFit()
-                                                        .frame(width: 40, height: 40)
-                                                }
-                                            }
-                                    }
-                                    .offset(x: 46, y: -40)
-                                    .sheet(isPresented: $isShowingMessageView) {
-                                        PostitView(postitVM: postitVM,
-                                                   isShowingMessageView: $isShowingMessageView)
-                                        .presentationDragIndicator(.visible)
-                                    }
-                                    .presentationDetents([.large])
-                                }
-                            
-                            
-                            HStack(alignment: .center, spacing: 4) {
-                                VStack(alignment: .leading, spacing: 4) {
-                                    if !user.username.isEmpty {
-                                        Text(user.username)
-                                            .font(.system(.body))
-                                            .fontWeight(.semibold)
-                                    } else {
-                                        Text(user.nameID)
-                                            .font(.system(.body))
-                                            .fontWeight(.semibold)
-                                    }
-                                }
-                                
-                                Spacer()
-                                
-                                if user.nameID == userNameID {
-                                    Button {
-                                        Task {
-                                            dismiss()
-                                            try? await Task.sleep(nanoseconds: 1 * 500_000_000)
-                                            viewModel.showTab = 4
-                                        }
-                                    } label: {
-                                        ZStack {
-                                            RoundedRectangle(cornerRadius:4)
-                                                .stroke(Color.white, lineWidth: 1)
-                                                .frame(width: 80, height: 28)
-                                            Text("내 프로필")
-                                                .font(.system(.footnote))
-                                                .fontWeight(.semibold)
-                                                .foregroundStyle(.white)
-                                        }
-                                    }
-                                } else {
-                                    FollowButtonView(buttonActive: $buttonOnOff, 
-                                                     cellUser: user,
-                                                     activeText: "팔로우",
-                                                     unActiveText: "팔로우 취소",
-                                                     widthValue: 85,
-                                                     heightValue: 28,
-                                                     buttonType: ButtonType.direct)
-                                }
-                            }
-                            
-                            HStack {
-                                HStack(spacing: 2) {
-                                    Text("\(profileVM.padoPosts.count + profileVM.sendPadoPosts.count)")
-                                    
-                                    Text("파도")
-                                        .fontWeight(.medium)
-                                }
-                                .font(.system(.body))
-                                .foregroundStyle(.white)
-                                
-                                Button {
-                                    followerActive = true
-                                } label: {
-                                    HStack(spacing: 2) {
-                                        Text("\(followVM.followerIDs.count + followVM.surferIDs.count)")
-                                        
-                                        Text("팔로워")
-                                            .fontWeight(.medium)
-                                    }
-                                    .font(.system(.body))
-                                    .foregroundStyle(.white)
-                                }
-                                .sheet(isPresented: $followerActive) {
-                                    FollowMainView(followVM: followVM, 
-                                                   currentType: "팔로워",
-                                                   user: user)
-                                        .presentationDetents([.large])
-                                        .presentationDragIndicator(.visible)
-                                        .onDisappear {
-                                            followerActive = false
-                                        }
-                                }
-                                
-                                Button {
-                                    followingActive = true
-                                } label: {
-                                    HStack(spacing: 2) {
-                                        Text("\(followVM.followingIDs.count)")
-                                        
-                                        Text("팔로잉")
-                                            .fontWeight(.medium)
-                                    }
-                                    .font(.system(.body))
-                                    .foregroundStyle(.white)
-                                }
-                                
-                                .sheet(isPresented: $followingActive) {
-                                    FollowMainView(followVM: followVM, 
-                                                   currentType: "팔로잉",
-                                                   user: user)
-                                    .presentationDetents([.large])
-                                    .presentationDragIndicator(.visible)
-                                    .onDisappear {
-                                        followingActive = false
-                                    }
-                                }
-                            }
-                            .padding(.leading, 2)
-                        }
-                        .padding(.horizontal)
-                        .padding(.bottom, 20)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                    }
-                }
-                .cornerRadius(0)
-                .offset(y: -minY)
-                .zIndex(touchBackImage ? 1 : 0)
-                .onTapGesture {
-                    if user.backProfileImageUrl != nil {
-                        position = .zero
-                        withAnimation(.interactiveSpring(response: 0.4, dampingFraction: 0.8, blendDuration: 0.8)) {
-                            touchBackImage = true
-                        }
-                    }
-                }
-        }
-        .frame(height: 300)
     }
     
     @ViewBuilder
