+++ PADO/PADO/iOS/Profile/View/ProfileView.swift
@@ -27,6 +27,7 @@ struct ProfileView: View {
     @State private var isDragging = false
     @State private var position = CGSize.zero
     @State private var isRefresh: Bool = false
+    @State private var buttonOnOff: Bool = false
     
     @State private var followerActive: Bool = false
     @State private var followingActive: Bool = false
@@ -44,8 +45,18 @@ struct ProfileView: View {
         NavigationStack {
             ZStack {
                 ScrollView(.vertical, showsIndicators: false) {
-                    headerView()
+                    ProfileHeaderView(touchBackImage: $touchBackImage,
+                                      touchProfileImage: $touchProfileImage,
+                                      followerActive: $followerActive,
+                                      followingActive: $followingActive,
+                                      position: $position,
+                                      buttonOnOff: $buttonOnOff, 
+                                      user: user,
+                                      profileVM: profileVM,
+                                      followVM: followVM,
+                                      postitVM: postitVM)
                         .padding(.top, 50)
+                    
                     VStack(spacing: 0) {
                         LazyVStack(pinnedViews: [.sectionHeaders]) {
                             Section {
@@ -160,184 +171,6 @@ struct ProfileView: View {
                 SettingProfileView()
             }
         }
-    }
-    
-    @ViewBuilder
-    func headerView() -> some View {
-        GeometryReader { proxy in
-            let minY = proxy.frame(in: .named("SCROLL")).minY
-            let size = proxy.size
-            let height = (size.height + minY)
-            
-            KFImage(URL(string: user.backProfileImageUrl ?? defaultBackgroundImageUrl))
-                .resizable()
-                .scaledToFill()
-                .frame(width: size.width, height: height > 0 ? height : 0, alignment: .top)
-                .overlay {
-                    ZStack(alignment: .bottom) {
-                        LinearGradient(colors: [.clear,
-                                                .main.opacity(0.1),
-                                                .main.opacity(0.3),
-                                                .main.opacity(0.5),
-                                                .main.opacity(0.8),
-                                                .main.opacity(1)],
-                                       startPoint: .top,
-                                       endPoint: .bottom)
-                        
-                        VStack(alignment: .leading, spacing: 10) {
-                            CircularImageView(size: .xxxLarge, user: user)
-                                .zIndex(touchProfileImage ? 1 : 0)
-                                .onTapGesture {
-                                    if user.profileImageUrl != nil {
-                                        position = .zero
-                                        withAnimation(.interactiveSpring(response: 0.4, dampingFraction: 0.8, blendDuration: 0.8)) {
-                                            touchProfileImage = true
-                                        }
-                                    }
-                                }
-                                .overlay {
-                                    Button {
-                                        viewModel.isShowingMessageView = true
-                                    } label: {
-                                        Circle()
-                                            .frame(width: 30)
-                                            .foregroundStyle(.clear)
-                                            .overlay {
-                                                if postitVM.messages.isEmpty {
-                                                    LottieView(animation: .named("nonePostit"))
-                                                        .paused(at: .progress(1))
-                                                        .resizable()
-                                                        .scaledToFit()
-                                                        .frame(width: 40, height: 40)
-                                                } else {
-                                                    LottieView(animation: .named("Postit"))
-                                                        .looping()
-                                                        .resizable()
-                                                        .scaledToFit()
-                                                        .frame(width: 40, height: 40)
-                                                }
-                                            }
-                                    }
-                                    .offset(x: 46, y: -40)
-                                    .sheet(isPresented: $viewModel.isShowingMessageView) {
-                                        PostitView(postitVM: postitVM,
-                                                   isShowingMessageView: $viewModel.isShowingMessageView)
-                                        .presentationDragIndicator(.visible)
-                                    }
-                                    .presentationDetents([.large])
-                                }
-                            
-                            HStack(alignment: .center, spacing: 4) {
-                                VStack(alignment: .leading, spacing: 4) {
-                                    if let user = viewModel.currentUser {
-                                        if !user.username.isEmpty {
-                                            Text(user.username)
-                                                .font(.system(.body))
-                                                .fontWeight(.semibold)
-                                        } else {
-                                            Text(userNameID)
-                                                .font(.system(.body))
-                                                .fontWeight(.semibold)
-                                        }
-                                    }
-                                }
-                                
-                                Spacer()
-                                
-                                Button {
-                                    viewModel.showingProfileView = true
-                                } label: {
-                                    ZStack {
-                                        RoundedRectangle(cornerRadius: 4)
-                                            .stroke(Color.white, lineWidth: 1)
-                                            .frame(width: 80, height: 28)
-                                        Text("프로필 편집")
-                                            .font(.system(.footnote))
-                                            .fontWeight(.semibold)
-                                            .foregroundStyle(.white)
-                                    }
-                                }
-                            }
-                            
-                            HStack {
-                                HStack(spacing: 2) {
-                                    Text("\(profileVM.padoPosts.count + profileVM.sendPadoPosts.count)")
-                                    
-                                    Text("파도")
-                                        .fontWeight(.medium)
-                                }
-                                .font(.system(.body))
-                                .foregroundStyle(.white)
-                                
-                                Button {
-                                    followerActive = true
-                                } label: {
-                                    HStack(spacing: 2) {
-                                        Text("\(followVM.followerIDs.count + followVM.surferIDs.count)")
-                                        
-                                        Text("팔로워")
-                                            .fontWeight(.medium)
-                                    }
-                                    .font(.system(.body))
-                                    .foregroundStyle(.white)
-                                }
-                                .sheet(isPresented: $followerActive) {
-                                    FollowMainView(followVM: followVM, 
-                                                   currentType: "팔로워",
-                                                   user: user)
-                                    .presentationDetents([.large])
-                                    .presentationDragIndicator(.visible)
-                                    .onDisappear {
-                                        followerActive = false
-                                    }
-                                }
-                                
-                                Button {
-                                    followingActive = true
-                                } label: {
-                                    HStack(spacing: 2) {
-                                        Text("\(followVM.followingIDs.count)")
-                                        
-                                        Text("팔로잉")
-                                            .fontWeight(.medium)
-                                    }
-                                    .font(.system(.body))
-                                    .foregroundStyle(.white)
-                                }
-                                
-                                .sheet(isPresented: $followingActive) {
-                                    FollowMainView(followVM: followVM, 
-                                                   currentType: "팔로잉",
-                                                   user: user)
-                                    .presentationDetents([.large])
-                                    .presentationDragIndicator(.visible)
-                                    .onDisappear {
-                                        followingActive = false
-                                    }
-                                }
-                                
-                            }
-                            .padding(.leading, 2)
-                        }
-                        .padding(.horizontal)
-                        .padding(.bottom, 20)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                    }
-                }
-                .cornerRadius(0)
-                .offset(y: -minY)
-                .zIndex(touchBackImage ? 1 : 0)
-                .onTapGesture {
-                    if user.backProfileImageUrl != nil {
-                        position = .zero
-                        withAnimation(.interactiveSpring(response: 0.4, dampingFraction: 0.8, blendDuration: 0.8)) {
-                            touchBackImage = true
-                        }
-                    }
-                }
-            
-        }
-        .frame(height: 300)
     }
     
     @ViewBuilder
